+++ src/server/trpc/router/_app.ts
@@ -1,6 +1,4 @@
-
 import { coursesRouter } from './courses';
-import { creditsRouter } from './credits';
 import { planRouter } from './plan';
 import { templateRouter } from './template';
 import { userRouter } from './user';
@@ -11,7 +9,6 @@ export const appRouter = router({
   user: userRouter,
   plan: planRouter,
   template: templateRouter,
-  credits: creditsRouter,
   courses: coursesRouter,
   validator: validatorRouter,
 });
